+++ packages/tui/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.51.1] - 2026-02-02
 
 ### Added